requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
cryptography>=41.0.7
openai>=1.3.7
google-generativeai>=0.3.1
httpx>=0.25.2
tenacity>=8.2.3
structlog>=24.1.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
hypothesis>=6.92.1
requests>=2.31.0
